requirements.txt
anchorpy==0.15.0
anchorpy-core==0.1.2
anyio==3.6.2
asttokens==2.2.1
attrs==22.2.0
autoflake==1.7.8
autopep8==2.0.1
backcall==0.2.0
base58==2.1.1
based58==0.1.1
black==22.12.0
borsh-construct==0.1.0
cachetools==4.2.4
certifi==2022.12.7
click==8.1.3
construct==2.10.68
construct-typing==0.5.5
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
genpy==2021.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
ipython==8.9.0
jedi==0.18.2
jsonalias==0.1.1
jsonrpcclient==4.0.2
matplotlib-inline==0.1.6
mnemonic==0.20
more-itertools==8.14.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pyflakes==2.5.0
Pygments==2.14.0
pyheck==0.1.5
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-xprocess==0.18.1
python-dotenv==0.21.1
pytools==2022.1.14
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
solana==0.29.1
solders==0.14.3
stack-data==0.6.2
sumtypes==0.1a6
tomli==2.0.1
toolz==0.11.2
traitlets==5.9.0
typer==0.4.1
types-cachetools==4.2.10
typing-extensions==4.4.0
wcwidth==0.2.6
websockets==10.4
zstandard==0.17.0
